+++ nextjs/components.json
@@ -15,7 +15,8 @@
     "utils": "@/lib/utils",
     "ui": "@/components/ui",
     "lib": "@/lib",
-    "hooks": "@/hooks"
+    "hooks": "@/hooks",
+    "apiUtils": "@/app/api/utils"
   },
   "iconLibrary": "lucide"
 }
