--- nextjs/src/ai/dev.ts
@@ -1,3 +1,0 @@
-import { config } from 'dotenv';
-config();
-
