+++ nextjs/src/app/(app)/teams/[teamId]/cards-tab.tsx
@@ -9,7 +9,7 @@
  * are view-only.
  */
 
-import { useMemo } from "react";
+import { useCallback, useEffect, useMemo, useState } from "react";
 import { BingoCard } from "@/components/bingo-card";
 import { TeamMembersProvider } from "@/components/team-members-context";
 import { TeamWsProvider } from "@/components/team-ws-provider";
@@ -17,17 +17,82 @@ import { Avatar, AvatarFallback } from "@/components/ui/avatar";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Separator } from "@/components/ui/separator";
 import Link from "next/link";
+import { useToast } from "@/hooks/use-toast";
 import type { Team, BingoCardData } from "./types";
 
 interface CardsTabProps {
+  teamId: string;
   team: Team;
-  cards: BingoCardData[];
+  initialCards: BingoCardData[];
   currentUserId: string;
-  onCellUpdate: (cellId: string, newState: "pending" | "completed") => void;
-  onRefresh: () => void;
 }
 
-export function CardsTab({ team, cards, currentUserId, onCellUpdate, onRefresh }: CardsTabProps) {
+export function CardsTab({
+  teamId,
+  team,
+  initialCards,
+  currentUserId,
+}: CardsTabProps) {
+  const { toast } = useToast();
+  const [cards, setCards] = useState<BingoCardData[]>(initialCards);
+
+  const loadCards = useCallback(async () => {
+    try {
+      const cardsRes = await fetch(`/api/teams/${teamId}/cards`);
+      const cardsData = await cardsRes.json();
+      if (cardsRes.ok) {
+        setCards(cardsData.cards || []);
+      }
+    } catch {
+      toast({
+        title: "Error",
+        description: "Failed to refresh cards",
+        variant: "destructive",
+      });
+    }
+  }, [teamId, toast]);
+
+  useEffect(() => {
+    // When the game transitions to started after initial render, fetch cards.
+    if (team.status === "started" && cards.length === 0) {
+      loadCards();
+    }
+  }, [cards.length, loadCards, team.status]);
+
+  const handleCellUpdate = useCallback(
+    async (cellId: string, newState: "pending" | "completed") => {
+      try {
+        const response =
+          newState === "pending"
+            ? await fetch(`/api/cells/${cellId}/undo-complete`, { method: "POST" })
+            : await fetch(`/api/cells/${cellId}`, {
+                method: "PUT",
+                headers: { "Content-Type": "application/json" },
+                body: JSON.stringify({ state: newState }),
+              });
+
+        if (response.ok) {
+          await loadCards();
+          return;
+        }
+
+        const data = await response.json();
+        toast({
+          title: "Error",
+          description: data.error || "Failed to update cell",
+          variant: "destructive",
+        });
+      } catch {
+        toast({
+          title: "Error",
+          description: "An error occurred",
+          variant: "destructive",
+        });
+      }
+    },
+    [loadCards, toast]
+  );
+
   // Build the userId → username map once from the already-loaded team data.
   const membersMap = useMemo(
     () =>
@@ -62,7 +127,7 @@ export function CardsTab({ team, cards, currentUserId, onCellUpdate, onRefresh }
   }
 
   return (
-    <TeamWsProvider teamId={team.id} onRefresh={onRefresh}>
+    <TeamWsProvider teamId={team.id} onRefresh={loadCards}>
     <TeamMembersProvider members={membersMap}>
     <div className="space-y-8">
       <h2 className="text-2xl font-bold font-headline">Bingo Cards</h2>
@@ -97,8 +162,8 @@ export function CardsTab({ team, cards, currentUserId, onCellUpdate, onRefresh }
                     isOwner={isCurrentUser}
                     teamId={team.id}
                     currentUserId={currentUserId}
-                    onCellUpdate={isCurrentUser ? onCellUpdate : undefined}
-                    onRefresh={onRefresh}
+                    onCellUpdate={isCurrentUser ? handleCellUpdate : undefined}
+                    onRefresh={loadCards}
                   />
                 ) : (
                   <p className="text-muted-foreground text-center py-8">
