+++ nextjs/src/app/api/cells/[cellId]/edit/route.ts
@@ -2,84 +2,52 @@
  * Cell Edit API
  * Spec Reference: 09-bingo-card-editing.md
  */
-
-import { getCurrentUser } from "@/lib/auth";
 import { updateCellContent } from "@/lib/db";
 import { NextRequest, NextResponse } from "next/server";
 import { ResolutionType } from "@/lib/shared/types";
-
+import { errorResponse, withAuth, AuthContext } from "@/app/api/utils";
 
 /**
  * PUT /api/cells/[cellId]/edit - Edit a cell's content
  * Spec: 09-bingo-card-editing.md - Persisting a Cell Edit
  */
-export async function PUT(
+export const PUT = withAuth(async (
   request: NextRequest,
-  { params }: { params: Promise<{ cellId: string }> }
-) {
-  try {
-    const currentUser = await getCurrentUser();
+  { params, currentUser }: AuthContext<{ cellId: string }>
+) => {
+  const { cellId } = await params;
+  const body = await request.json().catch(() => ({}));
 
-    if (!currentUser) {
-      return NextResponse.json(
-        { error: 'Authentication required' },
-        { status: 401 }
-      );
-    }
+  const sourceType = body?.sourceType;
+  const sourceUserId = typeof body?.sourceUserId === 'string' ? body.sourceUserId : null;
+  const resolutionId = typeof body?.resolutionId === 'string' ? body.resolutionId : null;
+  const resolutionType = typeof body?.resolutionType === 'string'
+    && [ResolutionType.BASE, ResolutionType.COMPOUND, ResolutionType.ITERATIVE].includes(body.resolutionType as ResolutionType)
+    ? (body.resolutionType as ResolutionType)
+    : ResolutionType.BASE;
 
-    const { cellId } = await params;
-    const body = await request.json().catch(() => ({}));
-
-    const sourceType = body?.sourceType;
-    const sourceUserId = typeof body?.sourceUserId === 'string' ? body.sourceUserId : null;
-    const resolutionId = typeof body?.resolutionId === 'string' ? body.resolutionId : null;
-    const resolutionType = typeof body?.resolutionType === 'string'
-      && [ResolutionType.BASE, ResolutionType.COMPOUND, ResolutionType.ITERATIVE].includes(body.resolutionType as ResolutionType)
-      ? (body.resolutionType as ResolutionType)
-      : ResolutionType.BASE;
-
-    const validSourceTypes = ['team', 'member_provided', 'personal', 'empty'] as const;
-    if (!validSourceTypes.includes(sourceType)) {
-      return NextResponse.json(
-        { error: 'Invalid sourceType' },
-        { status: 400 }
-      );
-    }
-
-    if (sourceType === 'personal' && !resolutionId) {
-      return NextResponse.json(
-        { error: 'resolutionId is required for personal cells' },
-        { status: 400 }
-      );
-    }
-
-    if (sourceType === 'member_provided' && !resolutionId) {
-      return NextResponse.json(
-        { error: 'resolutionId is required for member_provided cells' },
-        { status: 400 }
-      );
-    }
-
-    const result = await updateCellContent(cellId, currentUser.id, {
-      resolutionId,
-      resolutionType,
-      sourceType,
-      sourceUserId,
-    });
-
-    if (!result.success) {
-      return NextResponse.json(
-        { error: result.error || 'Failed to update cell content' },
-        { status: 400 }
-      );
-    }
-
-    return NextResponse.json({ cell: result.cell });
-  } catch (error) {
-    console.error('Edit cell error:', error);
-    return NextResponse.json(
-      { error: 'An error occurred' },
-      { status: 500 }
-    );
+  const validSourceTypes = ['team', 'member_provided', 'personal', 'empty'] as const;
+  if (!validSourceTypes.includes(sourceType)) {
+    return errorResponse('Invalid sourceType', 400);
   }
-}
+
+  if (sourceType === 'personal' && !resolutionId) {
+    return errorResponse('resolutionId is required for personal cells', 400);
+  }
+
+  if (sourceType === 'member_provided' && !resolutionId) {
+    return errorResponse('resolutionId is required for member_provided cells', 400);
+  }
+  const result = await updateCellContent(cellId, currentUser.id, {
+    resolutionId,
+    resolutionType,
+    sourceType,
+    sourceUserId,
+  });
+
+  if (!result.success) {
+    return errorResponse(result.error || 'Failed to update cell content', 400);
+  }
+
+  return NextResponse.json({ cell: result.cell });
+});
