+++ nextjs/src/app/api/cells/[cellId]/proof/route.ts
@@ -4,18 +4,17 @@
  */
 
 import { NextRequest, NextResponse } from 'next/server';
-import { getCurrentUser } from '@/lib/auth';
 import { mkdir, writeFile } from 'node:fs/promises';
 import path from 'node:path';
-import { randomUUID } from 'node:crypto';
-import { 
-  uploadProof, 
-  getProofsForCell, 
+import {
+  uploadProof,
+  getProofsForCell,
   reviewProof,
   getProofById,
 } from '@/lib/db';
 import type { ReviewDecision } from '@/lib/db/types';
-import { convertToWebP, tryConvertToWebP } from '@/lib/uploads-processing';
+import { errorResponse, withAuth, AuthContext } from '@/app/api/utils';
+import { tryConvertToWebP } from '@/lib/uploads-processing';
 
 const MAX_PROOF_FILE_BYTES = 5 * 1024 * 1024; // 5MB
 const ALLOWED_MIME_TYPES = new Set<string>([
@@ -50,87 +49,43 @@ function getSafeExtension(file: File): string | null {
 /**
  * GET /api/cells/[cellId]/proof - Get proofs for a cell
  */
-export async function GET(
-  _request: NextRequest,
-  { params }: { params: Promise<{ cellId: string }> }
-) {
-  try {
-    const currentUser = await getCurrentUser();
-    
-    if (!currentUser) {
-      return NextResponse.json(
-        { error: 'Authentication required' },
-        { status: 401 }
-      );
-    }
-
-    const { cellId } = await params;
-
-    const proofs = await getProofsForCell(cellId);
-    
-    return NextResponse.json({ proofs });
-  } catch (error) {
-    console.error('Get proofs error:', error);
-    return NextResponse.json(
-      { error: 'An error occurred' },
-      { status: 500 }
-    );
-  }
-}
+export const GET = withAuth(async (_request: NextRequest, { params, currentUser }: AuthContext<{ cellId: string }>) => {
+  const { cellId } = await params;
+  const proofs = await getProofsForCell(cellId);
+  return NextResponse.json({ proofs });
+});
 
 /**
  * POST /api/cells/[cellId]/proof - Upload proof for a cell
  * Spec: 07-proof-and-approval.md - A user can attach proof to a specific resolution cell
  */
-export async function POST(
-  request: NextRequest,
-  { params }: { params: Promise<{ cellId: string }> }
-) {
-  try {
-    const currentUser = await getCurrentUser();
-    
-    if (!currentUser) {
-      return NextResponse.json(
-        { error: 'Authentication required' },
-        { status: 401 }
-      );
-    }
+export const POST = withAuth(async (request: NextRequest, { params, currentUser }: AuthContext<{ cellId: string }>) => {
+  const { cellId } = await params;
+  const contentType = request.headers.get('content-type') || '';
 
-    const { cellId } = await params;
-    const contentType = request.headers.get('content-type') || '';
+  let fileUrl: string | undefined;
+  let comment: string | undefined;
 
-    let fileUrl: string | undefined;
-    let comment: string | undefined;
+  if (contentType.includes('multipart/form-data')) {
+    const form = await request.formData();
+    const formComment = form.get('comment');
+    comment = typeof formComment === 'string' ? formComment : undefined;
 
-    if (contentType.includes('multipart/form-data')) {
-      const form = await request.formData();
-      const formComment = form.get('comment');
-      comment = typeof formComment === 'string' ? formComment : undefined;
+    const file = form.get('file');
+    if (file instanceof File) {
+      const ext = getSafeExtension(file);
+      if (!ext) {
+        return errorResponse('Unsupported file type', 400);
+      }
 
-      const file = form.get('file');
-      if (file instanceof File) {
-        const ext = getSafeExtension(file);
-        if (!ext) {
-          return NextResponse.json(
-            { error: 'Unsupported file type' },
-            { status: 400 }
-          );
-        }
+      // If file.type is available, enforce it.
+      if (file.type && !ALLOWED_MIME_TYPES.has(file.type)) {
+        return errorResponse('Unsupported file type', 400);
+      }
 
-        // If file.type is available, enforce it.
-        if (file.type && !ALLOWED_MIME_TYPES.has(file.type)) {
-          return NextResponse.json(
-            { error: 'Unsupported file type' },
-            { status: 400 }
-          );
-        }
-
-        if (file.size > MAX_PROOF_FILE_BYTES) {
-          return NextResponse.json(
-            { error: 'File too large (max 5MB)' },
-            { status: 400 }
-          );
-        }
+      if (file.size > MAX_PROOF_FILE_BYTES) {
+        return errorResponse('File too large (max 5MB)', 400);
+      }
 
         const { buffer: convertedBuffer, type: convertedType, cadKey } = await tryConvertToWebP(file);
             
@@ -150,92 +105,53 @@ export async function POST(
       comment = body?.comment;
     }
 
-    // Upload proof
-    // Spec: 07-proof-and-approval.md - Only the card owner can upload proof
-    const result = await uploadProof(cellId, currentUser.id, fileUrl, comment);
-    
-    if (!result.success) {
-      return NextResponse.json(
-        { error: result.error },
-        { status: 400 }
-      );
-    }
-    
-    return NextResponse.json({ proof: result.proof }, { status: 201 });
-  } catch (error) {
-    console.error('Upload proof error:', error);
-    return NextResponse.json(
-      { error: 'An error occurred' },
-      { status: 500 }
-    );
+  // Upload proof
+  // Spec: 07-proof-and-approval.md - Only the card owner can upload proof
+  const result = await uploadProof(cellId, currentUser.id, fileUrl, comment);
+
+  if (!result.success) {
+    return errorResponse(result.error, 400);
   }
-}
+
+  return NextResponse.json({ proof: result.proof }, { status: 201 });
+});
 
 /**
  * PUT /api/cells/[cellId]/proof - Review a proof (approve/decline)
  * Spec: 07-proof-and-approval.md - A reviewer can approve or decline with a comment
  */
-export async function PUT(
-  request: NextRequest,
-  { params }: { params: Promise<{ cellId: string }> }
-) {
-  try {
-    const currentUser = await getCurrentUser();
-    
-    if (!currentUser) {
-      return NextResponse.json(
-        { error: 'Authentication required' },
-        { status: 401 }
-      );
-    }
+export const PUT = withAuth(async (request: NextRequest, { params, currentUser }: AuthContext<{ cellId: string }>) => {
+  const { cellId } = await params;
+  const body = await request.json();
+  const { proofId, decision, comment } = body;
 
-    const { cellId } = await params;
-    const body = await request.json();
-    const { proofId, decision, comment } = body;
-
-    if (!proofId) {
-      return NextResponse.json(
-        { error: 'Proof ID is required' },
-        { status: 400 }
-      );
-    }
-
-    // Validate decision
-    if (!decision || !['approved', 'declined'].includes(decision)) {
-      return NextResponse.json(
-        { error: 'Decision must be "approved" or "declined"' },
-        { status: 400 }
-      );
-    }
-
-    // Review proof
-    // Spec: 07-proof-and-approval.md - Only team members (excluding owner) can approve/decline
-    const result = await reviewProof(
-      proofId, 
-      currentUser.id, 
-      decision as ReviewDecision, 
-      comment
-    );
-    
-    if (!result.success) {
-      return NextResponse.json(
-        { error: result.error },
-        { status: 400 }
-      );
-    }
-    
-    // Get updated proof
-    const updatedProof = await getProofById(proofId);
-    
-    return NextResponse.json({ 
-      review: result.review,
-      proof: updatedProof,
-    });
-  } catch (error) {
-    console.error('Review proof error:', error);
-    return NextResponse.json(
-      { error: 'An error occurred' },
-      { status: 500 }
-    );
+  if (!proofId) {
+    return errorResponse('Proof ID is required', 400);
   }
-}
+
+  // Validate decision
+  if (!decision || !['approved', 'declined'].includes(decision)) {
+    return errorResponse('Decision must be "approved" or "declined"', 400);
+  }
+
+  // Review proof
+  // Spec: 07-proof-and-approval.md - Only team members (excluding owner) can approve/decline
+  const result = await reviewProof(
+    proofId,
+    currentUser.id,
+    decision as ReviewDecision,
+    comment
+  );
+
+  if (!result.success) {
+    return errorResponse(result.error, 400);
+  }
+
+  // Get updated proof
+  const updatedProof = await getProofById(proofId);
+
+  return NextResponse.json({
+    review: result.review,
+    proof: updatedProof,
+  });
+});
