+++ nextjs/src/app/api/cells/[cellId]/request-proof/route.ts
@@ -4,44 +4,21 @@
  */
 
 import { NextRequest, NextResponse } from 'next/server';
-import { getCurrentUser } from '@/lib/auth';
-import { requestProof } from '@/lib/db';
+import { requestProof, User } from '@/lib/db';
+import { errorResponse, withAuth, AuthContext } from '@/app/api/utils';
 
 /**
  * POST /api/cells/[cellId]/request-proof - Request proof for a completed cell
  * Creates a review thread for the cell
  */
-export async function POST(
+export const POST = withAuth(async (
   _request: NextRequest,
-  { params }: { params: Promise<{ cellId: string }> }
-) {
-  try {
-    const currentUser = await getCurrentUser();
-    
-    if (!currentUser) {
-      return NextResponse.json(
-        { error: 'Authentication required' },
-        { status: 401 }
-      );
-    }
-
-    const { cellId } = await params;
-
-    const result = await requestProof(cellId, currentUser.id);
-    
-    if (!result.success) {
-      return NextResponse.json(
-        { error: result.error },
-        { status: 400 }
-      );
-    }
-    
-    return NextResponse.json({ thread: result.thread }, { status: 201 });
-  } catch (error) {
-    console.error('Request proof error:', error);
-    return NextResponse.json(
-      { error: 'An error occurred' },
-      { status: 500 }
-    );
+  { params, currentUser }: AuthContext<{ cellId: string }>
+) => {
+  const { cellId } = await params;
+  const result = await requestProof(cellId, currentUser.id);
+  if (!result.success) {
+    return errorResponse(result.error, 400);
   }
-}
+  return NextResponse.json({ thread: result.thread }, { status: 201 });
+});
