+++ nextjs/src/app/api/resolutions/all/route.ts
@@ -5,9 +5,9 @@
  */
 
 import { NextResponse } from 'next/server';
-import { getCurrentUser } from '@/lib/auth';
 import { getResolutionsByUser } from '@/lib/db';
 import type { Subtask } from '@/lib/shared/types';
+import { errorResponse, withAuth, AuthContextNoParams } from '../../utils';
 
 interface UnifiedResolution {
   id: string;
@@ -25,45 +25,35 @@ interface UnifiedResolution {
 /**
  * GET /api/resolutions/all — returns all personal resolutions of all types
  */
-export async function GET() {
-  try {
-    const currentUser = await getCurrentUser();
-    if (!currentUser) {
-      return NextResponse.json({ error: 'Authentication required' }, { status: 401 });
+export const GET = withAuth(async (request: Request, { currentUser }: AuthContextNoParams) => {
+  const allResolutions = await getResolutionsByUser(currentUser.id);
+
+  const resolutions: UnifiedResolution[] = allResolutions.map((r) => {
+    const base = {
+      id: r.id,
+      type: r.resolutionType as 'base' | 'compound' | 'iterative',
+      ownerUserId: r.ownerUserId,
+      title: r.title,
+      text: r.description ?? '',
+      createdAt: r.createdAt,
+      updatedAt: r.updatedAt,
+    };
+
+    if (r.resolutionType === 'compound') {
+      return { ...base, subtasks: r.subtasks ?? undefined };
     }
+    if (r.resolutionType === 'iterative') {
+      return { ...base, numberOfRepetition: r.numberOfRepetition ?? undefined, completedTimes: r.completedTimes };
+    }
+    return base;
+  });
 
-    const allResolutions = await getResolutionsByUser(currentUser.id);
+  // Sort by creation date (newest first)
+  resolutions.sort((a, b) => {
+    const da = new Date(a.createdAt).getTime();
+    const db = new Date(b.createdAt).getTime();
+    return db - da;
+  });
 
-    const resolutions: UnifiedResolution[] = allResolutions.map((r) => {
-      const base = {
-        id: r.id,
-        type: r.resolutionType as 'base' | 'compound' | 'iterative',
-        ownerUserId: r.ownerUserId,
-        title: r.title,
-        text: r.description ?? '',
-        createdAt: r.createdAt,
-        updatedAt: r.updatedAt,
-      };
-
-      if (r.resolutionType === 'compound') {
-        return { ...base, subtasks: r.subtasks ?? undefined };
-      }
-      if (r.resolutionType === 'iterative') {
-        return { ...base, numberOfRepetition: r.numberOfRepetition ?? undefined, completedTimes: r.completedTimes };
-      }
-      return base;
-    });
-
-    // Sort by creation date (newest first)
-    resolutions.sort((a, b) => {
-      const da = new Date(a.createdAt).getTime();
-      const db = new Date(b.createdAt).getTime();
-      return db - da;
-    });
-
-    return NextResponse.json({ resolutions });
-  } catch (error) {
-    console.error('Get all resolutions error:', error);
-    return NextResponse.json({ error: 'An error occurred' }, { status: 500 });
-  }
-}
+  return NextResponse.json({ resolutions });
+});
