+++ nextjs/src/app/api/resolutions/compound/[id]/toggle/route.ts
@@ -6,9 +6,9 @@
  */
 
 import { NextRequest, NextResponse } from 'next/server';
-import { getCurrentUser } from '@/lib/auth';
-import { toggleCompoundSubtask, getCompoundResolutionById, autoTransitionCellState } from '@/lib/db';
+import { toggleCompoundSubtask, getCompoundResolutionById, autoTransitionCellState, User } from '@/lib/db';
 import { ResolutionType } from '@/lib/shared/types';
+import { errorResponse, withAuth, AuthContext } from '@/app/api/utils';
 
 /**
  * PATCH /api/resolutions/compound/[id]/toggle
@@ -16,49 +16,38 @@ import { ResolutionType } from '@/lib/shared/types';
  *
  * Toggles the completed state of the subtask at the given index.
  */
-export async function PATCH(
+export const PATCH = withAuth(async (
   request: NextRequest,
-  { params }: { params: Promise<{ id: string }> }
-) {
-  try {
-    const currentUser = await getCurrentUser();
-    if (!currentUser) {
-      return NextResponse.json({ error: 'Authentication required' }, { status: 401 });
-    }
+  { params, currentUser }: AuthContext<{ id: string }>
+) => {
+  const { id } = await params;
+  const body = await request.json();
+  const { subtaskIndex } = body;
 
-    const { id } = await params;
-    const body = await request.json();
-    const { subtaskIndex } = body;
-
-    if (typeof subtaskIndex !== 'number' || subtaskIndex < 0) {
-      return NextResponse.json({ error: 'Valid subtaskIndex is required' }, { status: 400 });
-    }
-
-    const existing = await getCompoundResolutionById(id);
-    if (!existing) {
-      return NextResponse.json({ error: 'Resolution not found' }, { status: 404 });
-    }
-    if (existing.ownerUserId !== currentUser.id) {
-      return NextResponse.json({ error: 'You can only modify your own resolutions' }, { status: 403 });
-    }
-
-    const resolution = await toggleCompoundSubtask(id, currentUser.id, subtaskIndex);
-    if (!resolution) {
-      return NextResponse.json({ error: 'Failed to toggle subtask' }, { status: 500 });
-    }
-
-    // Auto-transition bingo cells: all subtasks done → completed, otherwise → pending
-    const allDone = resolution.subtasks?.every((s) => s.completed) ?? false;
-    const updatedCells = await autoTransitionCellState(
-      id,
-      ResolutionType.COMPOUND,
-      allDone
-    );
-
-    return NextResponse.json({ resolution, updatedCells });
-  } catch (error) {
-    console.error('Toggle compound subtask error:', error);
-    const message = error instanceof Error ? error.message : 'An error occurred';
-    return NextResponse.json({ error: message }, { status: 500 });
+  if (typeof subtaskIndex !== 'number' || subtaskIndex < 0) {
+    return errorResponse('Valid subtaskIndex is required', 400);
   }
-}
+
+  const existing = await getCompoundResolutionById(id);
+  if (!existing) {
+    return errorResponse('Resolution not found', 404);
+  }
+  if (existing.ownerUserId !== currentUser.id) {
+    return errorResponse('You can only modify your own resolutions', 403);
+  }
+
+  const resolution = await toggleCompoundSubtask(id, currentUser.id, subtaskIndex);
+  if (!resolution) {
+    return errorResponse('Failed to toggle subtask', 500);
+  }
+
+  // Auto-transition bingo cells: all subtasks done → completed, otherwise → pending
+  const allDone = resolution.subtasks?.every((s) => s.completed) ?? false;
+  const updatedCells = await autoTransitionCellState(
+    id,
+    ResolutionType.COMPOUND,
+    allDone
+  );
+
+  return NextResponse.json({ resolution, updatedCells });
+});
