+++ nextjs/src/app/api/resolutions/compound/route.ts
@@ -9,180 +9,142 @@
  */
 
 import { NextRequest, NextResponse } from 'next/server';
-import { getCurrentUser } from '@/lib/auth';
 import {
   createCompoundResolution,
   getCompoundResolutionsByUser,
   getCompoundResolutionById,
   updateCompoundResolution,
   deleteCompoundResolution,
+  User,
 } from '@/lib/db';
 import type { Subtask } from '@/lib/shared/types';
+import { errorResponse, withAuth, AuthContextNoParams } from '@/app/api/utils';
 
 /**
  * GET /api/resolutions/compound - Get compound resolutions
  * Without ?id: returns all compound resolutions for the current user.
  * With ?id=...: returns a single compound resolution by ID (any authenticated user).
  */
-export async function GET(request: NextRequest) {
-  try {
-    const currentUser = await getCurrentUser();
-    if (!currentUser) {
-      return NextResponse.json({ error: 'Authentication required' }, { status: 401 });
-    }
-
+export const GET = withAuth(async (request: NextRequest, { currentUser }: AuthContextNoParams) => {
     const { searchParams } = new URL(request.url);
-    const id = searchParams.get('id');
+  const id = searchParams.get('id');
 
-    if (id) {
-      const resolution = await getCompoundResolutionById(id);
-      if (!resolution) {
-        return NextResponse.json({ error: 'Resolution not found' }, { status: 404 });
-      }
-      return NextResponse.json({ resolution });
+  if (id) {
+    const resolution = await getCompoundResolutionById(id);
+    if (!resolution) {
+      return errorResponse('Resolution not found', 404);
     }
-
-    const resolutions = await getCompoundResolutionsByUser(currentUser.id);
-    return NextResponse.json({ resolutions });
-  } catch (error) {
-    console.error('Get compound resolutions error:', error);
-    return NextResponse.json({ error: 'An error occurred' }, { status: 500 });
+    return NextResponse.json({ resolution });
   }
-}
+
+  const resolutions = await getCompoundResolutionsByUser(currentUser.id);
+  return NextResponse.json({ resolutions });
+});
 
 /**
  * POST /api/resolutions/compound - Create a new compound resolution
  * Body: { title: string, subtasks: Subtask[], description?: string }
  */
-export async function POST(request: NextRequest) {
-  try {
-    const currentUser = await getCurrentUser();
-    if (!currentUser) {
-      return NextResponse.json({ error: 'Authentication required' }, { status: 401 });
-    }
+export const POST = withAuth(async (request: NextRequest, { currentUser }: { currentUser: User }) => {
+  const body = await request.json();
+  const { title, subtasks, description } = body;
 
-    const body = await request.json();
-    const { title, subtasks, description } = body;
-
-    if (!title || typeof title !== 'string' || title.trim().length === 0) {
-      return NextResponse.json({ error: 'Title is required' }, { status: 400 });
-    }
-
-    if (!Array.isArray(subtasks) || subtasks.length === 0) {
-      return NextResponse.json({ error: 'At least one subtask is required' }, { status: 400 });
-    }
-
-    // Validate subtask shape
-    for (const st of subtasks) {
-      if (!st.title || typeof st.title !== 'string') {
-        return NextResponse.json({ error: 'Each subtask must have a title' }, { status: 400 });
-      }
-    }
-
-    const normalizedSubtasks: Subtask[] = subtasks.map((st: Subtask) => ({
-      title: st.title.trim(),
-      description: st.description?.trim() ?? '',
-      completed: st.completed ?? false,
-    }));
-
-    const resolution = await createCompoundResolution(
-      currentUser.id,
-      title,
-      normalizedSubtasks,
-      description ?? null
-    );
-
-    return NextResponse.json({ resolution }, { status: 201 });
-  } catch (error) {
-    console.error('Create compound resolution error:', error);
-    return NextResponse.json({ error: 'An error occurred' }, { status: 500 });
+  if (!title || typeof title !== 'string' || title.trim().length === 0) {
+    return errorResponse('Title is required', 400);
   }
-}
+
+  if (!Array.isArray(subtasks) || subtasks.length === 0) {
+    return errorResponse('At least one subtask is required', 400);
+  }
+
+  // Validate subtask shape
+  for (const st of subtasks) {
+    if (!st.title || typeof st.title !== 'string') {
+      return errorResponse('Each subtask must have a title', 400);
+    }
+  }
+
+  const normalizedSubtasks: Subtask[] = subtasks.map((st: Subtask) => ({
+    title: st.title.trim(),
+    description: st.description?.trim() ?? '',
+    completed: st.completed ?? false,
+  }));
+
+  const resolution = await createCompoundResolution(
+    currentUser.id,
+    title,
+    normalizedSubtasks,
+    description ?? null
+  );
+
+  return NextResponse.json({ resolution }, { status: 201 });
+});
 
 /**
  * PUT /api/resolutions/compound - Update a compound resolution
  * Body: { id: string, title?: string, description?: string, subtasks?: Subtask[] }
  */
-export async function PUT(request: NextRequest) {
-  try {
-    const currentUser = await getCurrentUser();
-    if (!currentUser) {
-      return NextResponse.json({ error: 'Authentication required' }, { status: 401 });
-    }
+export const PUT = withAuth(async (request: NextRequest, { currentUser }: { currentUser: User }) => {
+  const body = await request.json();
+  const { id, title, description, subtasks } = body;
 
-    const body = await request.json();
-    const { id, title, description, subtasks } = body;
-
-    if (!id) {
-      return NextResponse.json({ error: 'Resolution ID is required' }, { status: 400 });
-    }
-
-    const existing = await getCompoundResolutionById(id);
-    if (!existing) {
-      return NextResponse.json({ error: 'Resolution not found' }, { status: 404 });
-    }
-    if (existing.ownerUserId !== currentUser.id) {
-      return NextResponse.json({ error: 'You can only modify your own resolutions' }, { status: 403 });
-    }
-
-    const updates: { title?: string; description?: string | null; subtasks?: Subtask[] } = {};
-    if (title !== undefined) updates.title = title;
-    if (description !== undefined) updates.description = description;
-    if (subtasks !== undefined) {
-      if (!Array.isArray(subtasks) || subtasks.length === 0) {
-        return NextResponse.json({ error: 'At least one subtask is required' }, { status: 400 });
-      }
-      updates.subtasks = subtasks.map((st: Subtask) => ({
-        title: st.title.trim(),
-        description: st.description?.trim() ?? '',
-        completed: st.completed ?? false,
-      }));
-    }
-
-    const resolution = await updateCompoundResolution(id, currentUser.id, updates);
-    if (!resolution) {
-      return NextResponse.json({ error: 'Resolution not found' }, { status: 404 });
-    }
-
-    return NextResponse.json({ resolution });
-  } catch (error) {
-    console.error('Update compound resolution error:', error);
-    return NextResponse.json({ error: 'An error occurred' }, { status: 500 });
+  if (!id) {
+    return errorResponse('Resolution ID is required', 400);
   }
-}
+
+  const existing = await getCompoundResolutionById(id);
+  if (!existing) {
+    return errorResponse('Resolution not found', 404);
+  }
+  if (existing.ownerUserId !== currentUser.id) {
+    return errorResponse('You can only modify your own resolutions', 403);
+  }
+
+  const updates: { title?: string; description?: string | null; subtasks?: Subtask[] } = {};
+  if (title !== undefined) updates.title = title;
+  if (description !== undefined) updates.description = description;
+  if (subtasks !== undefined) {
+    if (!Array.isArray(subtasks) || subtasks.length === 0) {
+      return errorResponse('At least one subtask is required', 400);
+    }
+    updates.subtasks = subtasks.map((st: Subtask) => ({
+      title: st.title.trim(),
+      description: st.description?.trim() ?? '',
+      completed: st.completed ?? false,
+    }));
+  }
+
+  const resolution = await updateCompoundResolution(id, currentUser.id, updates);
+  if (!resolution) {
+    return errorResponse('Resolution not found', 404);
+  }
+
+  return NextResponse.json({ resolution });
+});
 
 /**
  * DELETE /api/resolutions/compound?id=... - Delete a compound resolution
  */
-export async function DELETE(request: NextRequest) {
-  try {
-    const currentUser = await getCurrentUser();
-    if (!currentUser) {
-      return NextResponse.json({ error: 'Authentication required' }, { status: 401 });
-    }
+export const DELETE = withAuth(async (request: NextRequest, { currentUser }: { currentUser: User }) => {
 
-    const { searchParams } = new URL(request.url);
-    const id = searchParams.get('id');
-    if (!id) {
-      return NextResponse.json({ error: 'Resolution ID is required' }, { status: 400 });
-    }
-
-    const existing = await getCompoundResolutionById(id);
-    if (!existing) {
-      return NextResponse.json({ error: 'Resolution not found' }, { status: 404 });
-    }
-    if (existing.ownerUserId !== currentUser.id) {
-      return NextResponse.json({ error: 'You can only delete your own resolutions' }, { status: 403 });
-    }
-
-    const deleted = await deleteCompoundResolution(id, currentUser.id);
-    if (!deleted) {
-      return NextResponse.json({ error: 'Failed to delete resolution' }, { status: 500 });
-    }
-
-    return NextResponse.json({ message: 'Compound resolution deleted successfully' });
-  } catch (error) {
-    console.error('Delete compound resolution error:', error);
-    return NextResponse.json({ error: 'An error occurred' }, { status: 500 });
+  const { searchParams } = new URL(request.url);
+  const id = searchParams.get('id');
+  if (!id) {
+    return errorResponse('Resolution ID is required', 400);
   }
-}
+
+  const existing = await getCompoundResolutionById(id);
+  if (!existing) {
+    return errorResponse('Resolution not found', 404);
+  }
+  if (existing.ownerUserId !== currentUser.id) {
+    return errorResponse('You can only delete your own resolutions', 403);
+  }
+
+  const deleted = await deleteCompoundResolution(id, currentUser.id);
+  if (!deleted) {
+    return errorResponse('Failed to delete resolution', 500);
+  }
+
+  return NextResponse.json({ message: 'Compound resolution deleted successfully' });
+});
