+++ nextjs/src/app/api/resolutions/iterative/[id]/increment/route.ts
@@ -7,41 +7,34 @@
  */
 
 import { NextRequest, NextResponse } from 'next/server';
-import { getCurrentUser } from '@/lib/auth';
 import {
   getIterativeResolutionById,
   incrementIterativeResolution,
   autoTransitionCellState,
+  User,
 } from '@/lib/db';
 import { ResolutionType } from '@/lib/shared/types';
+import { AuthContext, errorResponse, withAuth } from '@/app/api/utils';
 
 /**
  * PATCH /api/resolutions/iterative/[id]/increment
  * Increments the completed_times counter by 1.
  */
-export async function PATCH(
-  _request: NextRequest,
-  { params }: { params: Promise<{ id: string }> }
-) {
-  try {
-    const currentUser = await getCurrentUser();
-    if (!currentUser) {
-      return NextResponse.json({ error: 'Authentication required' }, { status: 401 });
-    }
-
+export const PATCH = withAuth(
+  async (_request: NextRequest, { params, currentUser }: AuthContext<{ id: string }>) => {
     const { id } = await params;
 
     const existing = await getIterativeResolutionById(id);
     if (!existing) {
-      return NextResponse.json({ error: 'Resolution not found' }, { status: 404 });
+      return errorResponse('Resolution not found', 404);
     }
     if (existing.ownerUserId !== currentUser.id) {
-      return NextResponse.json({ error: 'You can only modify your own resolutions' }, { status: 403 });
+      return errorResponse('You can only modify your own resolutions', 403);
     }
 
     const resolution = await incrementIterativeResolution(id, currentUser.id);
     if (!resolution) {
-      return NextResponse.json({ error: 'Failed to increment' }, { status: 500 });
+      return errorResponse('Failed to increment', 500);
     }
 
     // Auto-transition bingo cells: threshold met → completed
@@ -53,8 +46,4 @@ export async function PATCH(
     );
 
     return NextResponse.json({ resolution, updatedCells });
-  } catch (error) {
-    console.error('Increment iterative resolution error:', error);
-    return NextResponse.json({ error: 'An error occurred' }, { status: 500 });
-  }
-}
+  });
