+++ nextjs/src/app/api/teams/[teamId]/cards/route.ts
@@ -4,60 +4,38 @@
  */
 
 import { NextRequest, NextResponse } from 'next/server';
-import { getCurrentUser } from '@/lib/auth';
-import { 
+import {
   isTeamMember,
   getTeamBingoCards,
   getBingoCard,
+  User,
 } from '@/lib/db';
+import { errorResponse, withAuth, AuthContext } from '@/app/api/utils';
 
 /**
  * GET /api/teams/[teamId]/cards - Get all bingo cards for the team
  * Spec: 08-visibility-and-updates.md - A team member can view bingo cards of other members
  */
-export async function GET(
-  request: NextRequest,
-  { params }: { params: Promise<{ teamId: string }> }
-) {
-  try {
-    const currentUser = await getCurrentUser();
-    
-    if (!currentUser) {
-      return NextResponse.json(
-        { error: 'Authentication required' },
-        { status: 401 }
-      );
-    }
+export const GET = withAuth(async (request: NextRequest, { params, currentUser }: AuthContext<{ teamId: string }>) => {
+  const { teamId } = await params;
+  const { searchParams } = new URL(request.url);
+  const userId = searchParams.get('userId');
 
-    const { teamId } = await params;
-    const { searchParams } = new URL(request.url);
-    const userId = searchParams.get('userId');
-
-    // Check if user is a member
-    // Spec: 08-visibility-and-updates.md - Only team members can view team cards
-    const isMember = await isTeamMember(teamId, currentUser.id);
-    if (!isMember) {
-      return NextResponse.json(
-        { error: 'You are not a member of this team' },
-        { status: 403 }
-      );
-    }
-
-    if (userId) {
-      // Get specific user's card
-      const card = await getBingoCard(teamId, userId);
-      return NextResponse.json({ card });
-    }
-
-    // Get all cards for the team
-    const cards = await getTeamBingoCards(teamId, currentUser.id);
-    
-    return NextResponse.json({ cards });
-  } catch (error) {
-    console.error('Get cards error:', error);
-    return NextResponse.json(
-      { error: 'An error occurred' },
-      { status: 500 }
-    );
+  // Check if user is a member
+  // Spec: 08-visibility-and-updates.md - Only team members can view team cards
+  const isMember = await isTeamMember(teamId, currentUser.id);
+  if (!isMember) {
+    return errorResponse('You are not a member of this team', 403);
   }
-}
+
+  if (userId) {
+    // Get specific user's card
+    const card = await getBingoCard(teamId, userId);
+    return NextResponse.json({ card });
+  }
+
+  // Get all cards for the team
+  const cards = await getTeamBingoCards(teamId, currentUser.id);
+
+  return NextResponse.json({ cards });
+});
