+++ nextjs/src/app/api/teams/[teamId]/leaderboard/route.ts
@@ -7,33 +7,23 @@
  */
 
 import { NextRequest, NextResponse } from "next/server";
-import { getCurrentUser } from "@/lib/auth";
-import { isTeamMember, getTeamLeaderboard } from "@/lib/db";
+import { isTeamMember, getTeamLeaderboard, User } from "@/lib/db";
+import { AuthContext, withAuth } from "@/app/api/utils";
 
 // ── Route handler ─────────────────────────────────────────────────
 
-export async function GET(
+export const GET = withAuth(async (
   _request: NextRequest,
-  { params }: { params: Promise<{ teamId: string }> }
-) {
-  try {
-    const currentUser = await getCurrentUser();
-    if (!currentUser) {
-      return NextResponse.json({ error: "Authentication required" }, { status: 401 });
-    }
+  { params, currentUser }: AuthContext<{ teamId: string }>
+) => {
+  const { teamId } = await params;
 
-    const { teamId } = await params;
-
-    const isMember = await isTeamMember(teamId, currentUser.id);
-    if (!isMember) {
-      return NextResponse.json({ error: "You are not a member of this team" }, { status: 403 });
-    }
-
-    const leaderboard = await getTeamLeaderboard(teamId);
-
-    return NextResponse.json({ leaderboard });
-  } catch (error) {
-    console.error("Leaderboard error:", error);
-    return NextResponse.json({ error: "An error occurred" }, { status: 500 });
+  const isMember = await isTeamMember(teamId, currentUser.id);
+  if (!isMember) {
+    return NextResponse.json({ error: "You are not a member of this team" }, { status: 403 });
   }
-}
+
+  const leaderboard = await getTeamLeaderboard(teamId);
+
+  return NextResponse.json({ leaderboard });
+});
