+++ nextjs/src/app/api/threads/[threadId]/files/route.ts
@@ -4,12 +4,11 @@
  */
 
 import { NextRequest, NextResponse } from 'next/server';
-import { getCurrentUser } from '@/lib/auth';
-import { deleteThreadFile, uploadFile } from '@/lib/db';
+import { deleteThreadFile, uploadFile, User } from '@/lib/db';
 import { mkdir, writeFile } from 'node:fs/promises';
 import path from 'node:path';
-import { randomUUID } from 'node:crypto';
-import { convertToWebP, tryConvertToWebP } from '@/lib/uploads-processing';
+import { errorResponse, withAuth, AuthContext } from '@/app/api/utils';
+import { tryConvertToWebP } from '@/lib/uploads-processing';
 
 const MAX_FILE_SIZE = 5 * 1024 * 1024; // 5MB
 const ALLOWED_MIME_TYPES = new Set<string>([
@@ -56,143 +55,85 @@ function getSafeExtension(file: File): string | null {
  * POST /api/threads/[threadId]/files - Upload a proof file to a thread
  * Only the completing user can upload files
  */
-export async function POST(
+export const POST = withAuth(async (
   request: NextRequest,
-  { params }: { params: Promise<{ threadId: string }> }
-) {
-  try {
-    const currentUser = await getCurrentUser();
-    
-    if (!currentUser) {
-      return NextResponse.json(
-        { error: 'Authentication required' },
-        { status: 401 }
-      );
-    }
+  { params, currentUser }: AuthContext<{ threadId: string }>
+) => {
+  const { threadId } = await params;
+  const contentType = request.headers.get('content-type') || '';
 
-    const { threadId } = await params;
-    const contentType = request.headers.get('content-type') || '';
-
-    if (!contentType.includes('multipart/form-data')) {
-      return NextResponse.json(
-        { error: 'Content-Type must be multipart/form-data' },
-        { status: 400 }
-      );
-    }
-
-    const form = await request.formData();
-    const file = form.get('file');
-
-    if (!(file instanceof File)) {
-      return NextResponse.json(
-        { error: 'File is required' },
-        { status: 400 }
-      );
-    }
-
-    // Validate file size
-    if (file.size > MAX_FILE_SIZE) {
-      return NextResponse.json(
-        { error: 'File size exceeds 5MB limit' },
-        { status: 400 }
-      );
-    }
-
-    // Validate file type
-    if (file.type && !ALLOWED_MIME_TYPES.has(file.type)) {
-      return NextResponse.json(
-        { error: 'Unsupported file type' },
-        { status: 400 }
-      );
-    }
-
-    const ext = getSafeExtension(file);
-    if (!ext) {
-      return NextResponse.json(
-        { error: 'Unsupported file type' },
-        { status: 400 }
-      );
-    }
-
-    // Save file to disk
-    const { buffer: convertedBuffer, type: convertedType , cadKey } = await tryConvertToWebP(file);
-    const uploadDir = path.join(process.cwd(), 'uploads', 'review-files');
-    await mkdir(uploadDir, { recursive: true });
-
-    const filePath = path.join(uploadDir, cadKey);
-    await writeFile(filePath, convertedBuffer);
-    const fileUrl = `/review-files/${cadKey}`;
-
-    // Save to database
-    const result = await uploadFile(
-      threadId,
-      currentUser.id,
-      fileUrl,
-      convertedBuffer.byteLength,
-      file.name,
-      convertedType
-    );
-    
-    if (!result.success) {
-      return NextResponse.json(
-        { error: result.error },
-        { status: 400 }
-      );
-    }
-    
-    return NextResponse.json({ file: result.file }, { status: 201 });
-  } catch (error) {
-    console.error('Upload file error:', error);
-    return NextResponse.json(
-      { error: 'An error occurred' },
-      { status: 500 }
-    );
+  if (!contentType.includes('multipart/form-data')) {
+    return errorResponse('Content-Type must be multipart/form-data', 400);
   }
-}
+
+  const form = await request.formData();
+  const file = form.get('file');
+
+  if (!(file instanceof File)) {
+    return errorResponse('File is required', 400);
+  }
+
+  // Validate file size
+  if (file.size > MAX_FILE_SIZE) {
+    return errorResponse('File size exceeds 5MB limit', 400);
+  }
+
+  // Validate file type
+  if (file.type && !ALLOWED_MIME_TYPES.has(file.type)) {
+    return errorResponse('Unsupported file type', 400);
+  }
+
+  const ext = getSafeExtension(file);
+  if (!ext) {
+    return errorResponse('Unsupported file type', 400);
+  }
+
+  // Save file to disk
+  const { buffer: convertedBuffer, type: convertedType, cadKey } = await tryConvertToWebP(file);
+  const uploadDir = path.join(process.cwd(), 'uploads', 'review-files');
+  await mkdir(uploadDir, { recursive: true });
+
+  const filePath = path.join(uploadDir, cadKey);
+  await writeFile(filePath, convertedBuffer);
+  const fileUrl = `/review-files/${cadKey}`;
+
+  // Save to database
+  const result = await uploadFile(
+    threadId,
+    currentUser.id,
+    fileUrl,
+    convertedBuffer.byteLength,
+    file.name,
+    convertedType
+  );
+
+  if (!result.success) {
+    return errorResponse(result.error, 400);
+  }
+
+  return NextResponse.json({ file: result.file }, { status: 201 });
+});
 
 /**
  * DELETE /api/threads/[threadId]/files - Delete a previously uploaded proof file
  * Only the completing user can delete files from an open thread
  */
-export async function DELETE(
+export const DELETE = withAuth(async (
   request: NextRequest,
-  { params }: { params: Promise<{ threadId: string }> }
-) {
-  try {
-    const currentUser = await getCurrentUser();
+  { params, currentUser }: AuthContext<{ threadId: string }>
+) => {
+  const { threadId } = await params;
+  const body = await request.json().catch(() => null);
+  const fileId = body?.fileId;
 
-    if (!currentUser) {
-      return NextResponse.json(
-        { error: 'Authentication required' },
-        { status: 401 }
-      );
-    }
-
-    const { threadId } = await params;
-    const body = await request.json().catch(() => null);
-    const fileId = body?.fileId;
-
-    if (!fileId || typeof fileId !== 'string') {
-      return NextResponse.json(
-        { error: 'File ID is required' },
-        { status: 400 }
-      );
-    }
-
-    const result = await deleteThreadFile(threadId, currentUser.id, fileId);
-    if (!result.success) {
-      return NextResponse.json(
-        { error: result.error },
-        { status: result.error === 'File not found' ? 404 : 400 }
-      );
-    }
-
-    return NextResponse.json({ success: true });
-  } catch (error) {
-    console.error('Delete file error:', error);
-    return NextResponse.json(
-      { error: 'An error occurred' },
-      { status: 500 }
-    );
+  if (!fileId || typeof fileId !== 'string') {
+    return errorResponse('File ID is required', 400);
   }
-}
+
+  const result = await deleteThreadFile(threadId, currentUser.id, fileId);
+  if (!result.success) {
+    return errorResponse(result.error, result.error === 'File not found' ? 404 : 400);
+  }
+
+  return NextResponse.json({ success: true });
+});
