+++ nextjs/src/app/api/threads/[threadId]/route.ts
@@ -4,81 +4,33 @@
  */
 
 import { NextRequest, NextResponse } from 'next/server';
-import { getCurrentUser } from '@/lib/auth';
-import { closeThread, getThreadById } from '@/lib/db';
+import { closeThread, getThreadById, User } from '@/lib/db';
+import { errorResponse, withAuth, AuthContext } from '@/app/api/utils';
 
 /**
  * GET /api/threads/[threadId] - Get thread details with messages, files, and votes
  */
-export async function GET(
-  _request: NextRequest,
-  { params }: { params: Promise<{ threadId: string }> }
-) {
-  try {
-    const currentUser = await getCurrentUser();
-    
-    if (!currentUser) {
-      return NextResponse.json(
-        { error: 'Authentication required' },
-        { status: 401 }
-      );
-    }
-
-    const { threadId } = await params;
-
-    const result = await getThreadById(threadId, currentUser.id);
-    
-    if (!result.success) {
-      return NextResponse.json(
-        { error: result.error },
-        { status: result.error === 'Thread not found' ? 404 : 403 }
-      );
-    }
-    
-    return NextResponse.json({ thread: result.thread });
-  } catch (error) {
-    console.error('Get thread error:', error);
-    return NextResponse.json(
-      { error: 'An error occurred' },
-      { status: 500 }
-    );
+export const GET = withAuth(async (request: NextRequest, { params, currentUser }: AuthContext<{ threadId: string }>) => {
+  const { threadId } = await params;
+  const result = await getThreadById(threadId, currentUser.id);
+  if (!result.success) {
+    return errorResponse(result.error!, result.error === 'Thread not found' ? 404 : 403);
   }
-}
+  return NextResponse.json({ thread: result.thread });
+});
 
 /**
  * DELETE /api/threads/[threadId] - Close a review thread
  * Only the resolution owner can close an open thread
  */
-export async function DELETE(
+export const DELETE = withAuth(async (
   _request: NextRequest,
-  { params }: { params: Promise<{ threadId: string }> }
-) {
-  try {
-    const currentUser = await getCurrentUser();
-
-    if (!currentUser) {
-      return NextResponse.json(
-        { error: 'Authentication required' },
-        { status: 401 }
-      );
-    }
-
-    const { threadId } = await params;
-    const result = await closeThread(threadId, currentUser.id);
-
-    if (!result.success) {
-      return NextResponse.json(
-        { error: result.error },
-        { status: result.error === 'Thread not found' ? 404 : 400 }
-      );
-    }
-
-    return NextResponse.json({ success: true });
-  } catch (error) {
-    console.error('Close thread error:', error);
-    return NextResponse.json(
-      { error: 'An error occurred' },
-      { status: 500 }
-    );
+  { params, currentUser }: AuthContext<{ threadId: string }>
+) => {
+  const { threadId } = await params;
+  const result = await closeThread(threadId, currentUser.id);
+  if (!result.success) {
+    return errorResponse(result.error!, result.error === 'Thread not found' ? 404 : 403);
   }
-}
+  return NextResponse.json({ success: true });
+});
