+++ nextjs/src/app/api/users/[id]/route.ts
@@ -1,29 +1,17 @@
 import type { NextRequest } from 'next/server';
 import { NextResponse } from 'next/server';
 import { findUserById } from '@/lib/db';
-import { getCurrentUser } from '@/lib/auth';
+import { errorResponse, withAuth, AuthContext } from '@/app/api/utils';
 
-export async function GET(
-  request: NextRequest,
-  { params }: { params: Promise<{ id: string }> }
-) {
-  try {
-    const currentUser = await getCurrentUser();
-        
-        if (!currentUser) {
-          return NextResponse.json(
-            { error: 'Authentication required' },
-            { status: 401 }
-          );
-        }
-    const { id } = await params;
-    if (!id) return NextResponse.json({ error: 'Missing id' }, { status: 400 });
+export const GET = withAuth(async (
+  _request: NextRequest,
+  { params }: AuthContext<{ id: string }>
+) => {
+  const { id } = await params;
+  if (!id) return errorResponse('Missing id', 400);
 
-    const user = await findUserById(id);
-    if (!user) return NextResponse.json({ error: 'User not found' }, { status: 404 });
+  const user = await findUserById(id);
+  if (!user) return errorResponse('User not found', 404);
 
-    return NextResponse.json({ id: user.id, username: user.username });
-  } catch (err) {
-    return NextResponse.json({ error: 'Server error' }, { status: 500 });
-  }
-}
+  return NextResponse.json({ id: user.id, username: user.username });
+});
