+++ nginx/nyr-bingo_nginx.conf
@@ -6,6 +6,16 @@ server {
 
   client_max_body_size 6m;
 
+  # gzip configuration
+  gzip on;
+  gzip_http_version 1.1;
+  gzip_comp_level 3;
+  gzip_min_length 1000;
+  gzip_types text/plain text/css application/json application/javascript text/javascript;
+  gzip_proxied any;
+  gzip_vary on; # Add Vary header to responses to indicate that the response may vary based on the Accept-Encoding header of the request
+  gzip_buffers 32 4k; # Exceeding the buffer, disk space will be used
+
   location /ws {
     proxy_pass http://ws-server:8888;
     proxy_http_version 1.1;
@@ -42,9 +52,18 @@ server {
     proxy_cache_valid 200 10m;
 
     # Source - https://stackoverflow.com/a/13659958
-    # Posted by ChocoDeveloper, modified by community. See post 'Timeline' for change history
+    # Posted by ChocoDeveloper, modified by community.
     # Retrieved 2026-01-15, License - CC BY-SA 3.0
     add_header X-Cache $upstream_cache_status;
     add_header Cache-Control "public, immutable"; # Client will not revalidate until the file is removed from cache
+
+    # Source - https://stackoverflow.com/questions/58066785/always-use-sendfile-with-nginx-on-linux
+    # Posted by: Hunter Kohler
+    # Retrieved 2026-03-29
+    sendfile           on;  # Use sendfile() --> avoid write and copy of data into socket buffer; directly send file
+    sendfile_max_chunk 1m;  # Limits chunks to 1 Megabytes; avoid slowing down the server when sending large files
+    tcp_nopush         on;  # Optimize amount of TCP packets sent by waiting until MSS is reached before sending. Requires sendfile set to 'on'
+    tcp_nodelay        on;  # Disable Nagle's algorithm - https://www.netguru.com/blog/nginx-tutorial-performance#:~:text=The%20Nagle%20algorithm%20was%20a%20true%20saviour%20in%20its%20time%20and%20still%20provides%20great%20value.%20However%2C%20in%20most%20cases%2C%20we%20won%E2%80%99t%20need%20it%20for%20our%20website%2C%20and%20it%20can%20be%20safely%20turned%20down%20via%20adding%20the%20flag%20TCP_NODELAY.
+                            # The use of sendfile, tcp_nopush and tcp_nodelay  ensures (i) packages are full before sending them and (ii) the last packet will not be delayed.
   }
 }
